requirements/dev.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.2
psycopg2-binary==2.9.1
python-decouple==3.4
whitenoise==5.3.0
